implement inherition
